feat(charm): add support for EFS volumes with tls
This commit adds the support for a new mount type called `efs`. This
mount type is specifically used for mounting AWS EFS volumes with support
for TLS encryption. This is achieved by adding a configuration option
called `mount_type` which can be set to `efs` to enable this feature.
This feature requires a new resource `amazon-efs-utils-deb` which should be a
a debian package used to install aws efs helper utils.